fixed bad cast in bar fixed up cpr a little re-arranged structs to prep for kids tweaked options wrote babysit function.

+++ bar/bar.c
@@ -900,7 +900,7 @@ int barread(bar_t *bar, int fd)
 			uint64_t *w64 = (uint64_t *)barbuf;
 			filebits = (word_t)w64[0];
 		} else {
-			uint64_t *w32 = (uint32_t *)barbuf;
+			uint32_t *w32 = (uint32_t *)barbuf;
 			filebits = (word_t)w32[0];
 		}
 	}
